Rename package to ebrains_validation_framework
