+++ MANIFEST.in
@@ -2,5 +2,5 @@ include setup.py
 include README.rst
 include LICENSE.txt
 include requirements.txt
-include hbp_validation_framework/*.py
-recursive-include hbp_validation_framework/jsonTreeViewer *
+include ebrains_validation_framework/*.py
+recursive-include ebrains_validation_framework/jsonTreeViewer *
