+++ docs/conf.py
@@ -19,7 +19,7 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath("./hbp_validation_framework"))
+sys.path.insert(0, os.path.abspath("./ebrains_validation_framework"))
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
